refactor(tool/cmd/importconfigs): move bazel package from internal/config
The bazel package is only used by the importconfigs tool. Move it from
internal/config/bazel to tool/cmd/importconfigs/bazel to colocate these
packages.
